update: file structure, sep app-server file
