+++ server/src/socket-server.ts
@@ -26,3 +26,5 @@ export class SocketServer {
     });
   }
 }
+
+export default SocketServer;
